# golang: speedup using TRIE