@starius 2024-04-13
Hi @javierron !

Thank you very much for the review!
I addressed the feedback, see my comments.

> Maybe the solution is to determine at initialization if a pattern is either literal or regex. I believe this would simplify analizePattern

Can you elaborate on this, please? `analizePattern` already determines automatically if a pattern is a literal:

```go
prefix, complete := re.LiteralPrefix()
if complete {
  return []string{prefix}, nil
}
```
(I put this code after checking presence in `pattern2literals` table, because it is faster than compiling the regexp, so regexp compilation is not done for cases from `pattern2literals` table.)

We still need both tables `pattern2literals` and `pattern2mainLiteral`, because we need to do something with non-literal regexps. Maybe `pattern2literals` could be generated automatically, but such a code would be quite complicated, so I think it is not worth it, to generate such a small table. And for `pattern2mainLiteral` I don't know how to automate its creation even in theory: we need to find a long enough literal in regexp which is present in any match. It requires in-depth analysis of the structure of regexp. And the hardcoded table is even smaller...

Do you have ideas how to do it in an elegant way?