@starius 2024-05-16
Hi @javierron @monperrus !

I think it is possible to implement `func analyzePattern` without hardcoded tables. Go has package [regexp/syntax](https://pkg.go.dev/regexp/syntax) which provides introspection of regular expressions. I think it is possible to use it to convert a pattern to the list of all possible matching strings (what `pattern2literals` does) and to extract longest common sub-string from a regexp (what `pattern2mainLiteral` does). I'll try to implement this in few days and update the PR.