@starius 2024-10-14
@monperrus @javierron  @srstsavage Please check the latest version of the PR!
The tables with hardcoded cases were replaced with `analyzePattern` function, which does it automatically.